fix: relocate debugger
