+++ perception/multi_object_tracker/CMakeLists.txt
@@ -22,7 +22,7 @@ include_directories(
 # Generate exe file
 set(MULTI_OBJECT_TRACKER_SRC
   src/multi_object_tracker_core.cpp
-  src/debugger.cpp
+  src/debugger/debugger.cpp
   src/processor/processor.cpp
   src/processor/input_manager.cpp
   src/data_association/data_association.cpp
