+++ perception/multi_object_tracker/include/multi_object_tracker/multi_object_tracker_core.hpp
@@ -20,7 +20,7 @@
 #define MULTI_OBJECT_TRACKER__MULTI_OBJECT_TRACKER_CORE_HPP_
 
 #include "multi_object_tracker/data_association/data_association.hpp"
-#include "multi_object_tracker/debugger.hpp"
+#include "multi_object_tracker/debugger/debugger.hpp"
 #include "multi_object_tracker/processor/input_manager.hpp"
 #include "multi_object_tracker/processor/processor.hpp"
 #include "multi_object_tracker/tracker/model/tracker_base.hpp"
